put a 1hr timeout on flock blocking
From time to time the flock calls on the er system back up for an
unknown reason. However, when fixing another bug I found that flock
has a wait timeout. Set this to an hour so that if we fail to flock
after an hour we don't back up forever.

Change-Id: Ie574b994d2b5cc38f559d3d7f8dd980148cf0890