Rework uncategorized script to handle multiple html files
In a recent elastic-recheck change we updated the output of the
uncategorized command to split between the integrated gate failues
and everything else. In the future we'll add a configurable views
system where we can construct multiple different groups. The days
of a single uncategorized.html file have disappeared. This commit
adds the necessary support to the cron job so it can handle an
multiple number of html files.

Change-Id: I12690cecd3b5d27fa3e508f7ea4daa43696423b8