Add beaker tests
Add tests to ensure the elastic_recheck classes work and that the
service can start.

This module hardcodes IRC bot configs, so the tests use resource
overrides to set the bot channels to a test channel that's also used
elsewhere in our tests to avoid spamming our real channels.

In order for the bot to actually be functional it would need to have a
real gerrit account, which we don't have, but this should still be
sufficient as a test.

Change-Id: I6da62d63d885037ecbc1560deec323e0d8583c3c